Add check for value size before sending to memcached
This is needed because memcached, by default, cannot store objects
greater than 1MB in size. However, this is configurable. This change
adds a configuration parameter to specify a max_value_size and error if
it is exceeded.

As a side effect, the error message for a 0 length key has been changed
to MEMCACHED_NO_KEY_PROVIDED instead of MEMCACHED_NOTSTORED because of the
error reporting collision introduced by this change.